+++ assets/javascripts/kitten/components/action/button/index.js
@@ -241,9 +241,9 @@ export const Button = ({
       className={classNames(
         'k-Button',
         className,
-        `k-Button--${size}`,
-        `k-Button--${internalModifier}`,
-        `k-Button--fit-${fit}`,
+        `k-Button--${size || 'regular'}`,
+        `k-Button--${internalModifier || 'hydrogen'}`,
+        `k-Button--fit-${fit || 'min-width'}`,
         {
           [`k-Button--mobile-fit-${mobileFit}`]: !!mobileFit,
           'k-Button--disabled': disabled,