+++ assets/javascripts/kitten/components/navigation/header-nav/stories/kisskissbankbank.js
@@ -271,6 +271,7 @@ export const KissKissBankBankHeaderNavStory = ({
               size: 'tiny',
               rounded: true,
               modifier: 'hydrogen',
+              fit: 'icon',
             }}
             className="k-u-hidden@m-down"
           >