+++ src/components/form/field/components/radio-set.js
@@ -7,12 +7,12 @@ exports.FieldRadioSet = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
-var _radioButtonSet = require("../../../form/radio-button-set");
+var _radioSet = require("../../../form/radio-set");
 
 var FieldRadioSet = function FieldRadioSet(props) {
   return /*#__PURE__*/_react.default.createElement("div", {
     className: "k-u-margin-top-single"
-  }, /*#__PURE__*/_react.default.createElement(_radioButtonSet.RadioSet, props));
+  }, /*#__PURE__*/_react.default.createElement(_radioSet.RadioSet, props));
 };
 
 exports.FieldRadioSet = FieldRadioSet;