+++ src/esm/components/form/field/components/radio-set.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RadioSet } from '../../../form/radio-button-set';
+import { RadioSet } from '../../../form/radio-set';
 export var FieldRadioSet = function FieldRadioSet(props) {
   return /*#__PURE__*/React.createElement("div", {
     className: "k-u-margin-top-single"