+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Deprecations/Changes:`ButtonIcon`: Deprecate component, `Button fit="icon"` benefit.
+- Breaking changes: `Tokens`: Update token color with semantic color.
 - Fix: `SummaryCard`: Improve `contributorSubscription` columns.
 
 ## [8.9.1] - 2022-03-16