+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `HeaderNav`: Fix horizontal scroll.
 - Deprecations/Changes:`ButtonIcon`: Deprecate component, `Button fit="icon"` benefit.
 - Deprecations/Changes: `Button`: Deprecate `fluid` and `icon` props.
 - Breaking changes: `Tokens`: Update token color with semantic color.