+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `Alert`: Improve styles.
 - Fix: `Button`: Force default values for props.
 
 ## [9.0.0] - 2022-03-21