+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add the following icon components:
   - `EmbedIconNext`
   - `LinkIconNext`
+- Fix: `Alert`: Improve styles.
+- Fix: `Button`: Force default values for props.
 
 ## [9.0.0] - 2022-03-21