+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: `Radio`: Add component.
 - Feature: `RadioSet`: Add component.
 - Feature: `Field`: Add `RadioSet` subcomponent.
+- Fix: `HeaderNav`: Fix overflow bug with `ButtonWithTooltip`.
 
 ## [9.1.0] - 2022-03-25