+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: `Field`: Fix `Field.RadioSet` import error.
+## [9.2.1] - 2022-03-28
+
+Fix:
+
+- `Field`: Fix `Field.RadioSet` import error.
 
 ## [9.2.0] - 2022-03-28